move service to separate class
